% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heritability.R
\name{heritability}
\alias{heritability}
\title{Heritability}
\usage{
heritability(genotypes, environments = NULL, outcome, rep,
  block = NULL, return = TRUE)
}
\arguments{
\item{genotypes}{vector containing genotypes}

\item{environments}{vector containing environments}

\item{outcome}{vector containing outcome}

\item{rep}{vector containing reps}

\item{block}{vector containing blocks}

\item{return}{TRUE/FALSE return the heritability stat in the console}
}
\description{
Heritability as calculated in a multi environment trial (as implemented in META-R)
}
\examples{
library(agridat)
data(besag.met)
with(besag.met,heritability(genotypes=gen,environments=county,outcome=yield,rep=rep,block=block))
}
\keyword{heritability}
