% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envcorr.R
\name{envcorr}
\alias{envcorr}
\title{Environmental Correlation}
\usage{
envcorr(genotypes, environments = NULL, outcomes, data, rep,
  block = NULL)
}
\arguments{
\item{genotypes}{vector containing genotypes}

\item{environments}{vector containing environments}

\item{outcomes}{vector of names of outcome variables to calculate correlation on}

\item{data}{dataframe containing outcomes}

\item{rep}{vector containing reps}

\item{block}{vector containing blocks}
}
\description{
Environmental Correlation (as implemented in META-R with error related to missing the reference level fixed )
}
\examples{
library(agridat)
data(yacon)
yaconF0<-subset(yacon,dose=="F0")
with(yaconF0,envcorr(genotypes=entry,environments=locality,outcomes=c("height","stalks","wfr","wff","wfk","roots","FOS","glucose",
"fructose","brix","foliage","dry","IH"),data=yaconF0,rep=replication))
}
\keyword{heritability}
