% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gencorr.R
\name{gencorr}
\alias{gencorr}
\title{Genotypic Correlation}
\usage{
gencorr(genotypes, environments = NULL, outcomes, data, rep,
  block = NULL, threshold = 0.05)
}
\arguments{
\item{genotypes}{vector containing genotypes}

\item{environments}{vector containing environments}

\item{outcomes}{vector of names of outcome variables to calculate correlation on}

\item{data}{dataframe containing outcomes}

\item{rep}{vector containing reps}

\item{block}{vector containing blocks}
}
\description{
Genotypic Correlation
}
\examples{
library(agricolae)
data(yacon)
yaconF150<-subset(yacon,dose=="F150")
with(yaconF150,gencorr(genotypes=entry,environments=locality,
outcomes=yaconF150[,c("height","stalks","wfr","wff","wfk","roots","FOS","glucose","fructose","brix","foliage","dry","IH")],
rep=replication)
}
\keyword{heritability}
